PR suggestions: properly commented the API. Move methods of serialization in private extensions in order to exclude it from the API.
