+++ lib/launchdarkly_flutter.dart
@@ -2,102 +2,10 @@ import 'dart:async';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
+import 'package:launchdarkly_flutter/launchdarkly_config.dart';
+import 'package:launchdarkly_flutter/launchdarkly_user.dart';
 
-class LaunchDarklyConfig {
-  final bool allAttributesPrivate;
-  final Set<String> privateAttributes;
-
-  const LaunchDarklyConfig({
-    this.allAttributesPrivate = false,
-    this.privateAttributes = const {},
-  });
-
-  Map<String, dynamic> toMap() => {
-        'allAttributesPrivate': allAttributesPrivate,
-        'privateAttributes': privateAttributes.toList(),
-      };
-}
-
-/// A collection of attributes that can affect flag evaluation, usually corresponding to a user of your application.
-///
-/// If you want to avoid sending personal information back to LaunchDarkly but keep the ability to target user segments, you can configure those attributes as private.
-/// Use `private`-prefixed counterparts for this purpose.
-class LaunchDarklyUser {
-  /// Sets the secondary key for a user. This affects feature flag targeting as follows:
-  /// if you have chosen to bucket users by a specific attribute,
-  /// he secondary key (if set) is used to further distinguish between users who are otherwise identical according to that attribute.
-  final String? secondaryKey;
-
-  /// Sets the IP for a user.
-  final String? ip;
-
-  /// Set the country for a user.
-  final String? country;
-
-  /// Sets the user's avatar.
-  final String? avatar;
-
-  /// Sets the user's e-mail address.
-  final String? email;
-
-  /// Sets the user's full name.
-  final String? name;
-
-  /// Sets the user's first name.
-  final String? firstName;
-
-  /// Sets the user's last name.
-  final String? lastName;
-
-  final List<String> privateAttributes;
-
-  LaunchDarklyUser({
-    String? secondaryKey,
-    String? privateSecondaryKey,
-    String? ip,
-    String? privateIp,
-    String? country,
-    String? privateCountry,
-    String? avatar,
-    String? privateAvatar,
-    String? email,
-    String? privateEmail,
-    String? name,
-    String? privateName,
-    String? firstName,
-    String? privateFirstName,
-    String? lastName,
-    String? privateLastName,
-  })  : this.secondaryKey = privateSecondaryKey ?? secondaryKey,
-        this.ip = privateIp ?? ip,
-        this.country = privateCountry ?? country,
-        this.avatar = privateAvatar ?? avatar,
-        this.email = privateEmail ?? email,
-        this.name = privateName ?? name,
-        this.firstName = privateFirstName ?? firstName,
-        this.lastName = privateLastName ?? lastName,
-        this.privateAttributes = [
-          if (privateSecondaryKey != null) 'secondary',
-          if (privateIp != null) 'ip',
-          if (privateCountry != null) 'country',
-          if (privateAvatar != null) 'avatar',
-          if (privateEmail != null) 'email',
-          if (privateName != null) 'name',
-          if (privateFirstName != null) 'firstName',
-          if (privateLastName != null) 'lastName',
-        ];
-
-  Map<String, dynamic> toMap() => {
-        'secondary': secondaryKey,
-        'ip': ip,
-        'country': country,
-        'avatar': avatar,
-        'email': email,
-        'name': name,
-        'firstName': firstName,
-        'lastName': lastName,
-      };
-}
+part 'launch_darkly_extensions.dart';
 
 /// Client for accessing LaunchDarkly's Feature Flag system.
 class LaunchdarklyFlutter {
@@ -188,6 +96,7 @@ class LaunchdarklyFlutter {
       return await _channel.invokeMethod('init', <String, dynamic>{
         'mobileKey': mobileKey,
         'config': config?.toMap(),
+        'user': user?.toMap(),
         'custom': {
           if (custom != null) ...custom,
           if (privateCustom != null) ...privateCustom,
